Add files via upload
Python script for calling C++ code  and writing the output to a file. Calls  C++ code running the Beta(2-a,a) coalescent or the Kingman-coalescent, simulates trees and  computes 
summary statistics and   writes the output to files.  Runs over a specified grid of parameter values (a,h), where a is the  alpha parameter of the Beta(2-a,a)-coalescent, or  the growth parameter of exponential population growth; and h is the scaled mutation rate (theta).   

The script is part of a manuscript currently undergoing peer-review.