Python script for simulating points
Python script for calling C++ code for  simulating points for the genomic  Kingman(beta)-coalescent, i.e.\ the  genomic time-changed Kingman-coalescent  representing  a diploid population growing according to the Wright-Fisher model and growing exponentially in size.      The output is then written to files.

The script is part  of a manuscript currently undergoing peer-review.  Citation details will follow in due course.